model/requirements.txt
google-generativeai>=0.1.0
flask>=2.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pydantic>=1.10.0
requests>=2.28.0
gunicorn>=21.2.0